pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "imery"
version = "0.0.3"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "imgui-bundle>=1.92.5",
    "munch>=4.0.0",
    "pillow>=12.0.0",
    "pyglm>=2.8.3",
    "pyside6>=6.10.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
    "stringcase>=1.2.0",
    "twine>=6.2.0",
    "wgpu>=0.28.1",
]

[project.scripts]
imery = "imery.app:main"

[tool.hatch.build.targets.wheel]
packages = ["src/imery"]

[tool.hatch.build.targets.wheel.force-include]
"demo" = "imery/demo"

[tool.hatch.build.targets.sdist]
include = [
    "src/imery/**/*.py",
    "src/imery/**/*.yaml",
    "demo/**/*",
]
